fix warnings for Clang

+++ src/Bpp/Seq/AlphabetIndex/CodonFromProteicAlphabetIndex1.h
@@ -115,8 +115,8 @@ namespace bpp
   private:
     void fillIndex_(const AlphabetIndex1* protAlphIndex_)
     {
-      for (int i=0; i<64; i++)
-        vIndex_[i]=gencode_->isStop(i)?0:protAlphIndex_->getIndex(gencode_->translate(i));
+      for (size_t i=0; i<64; i++)
+        vIndex_[i]=gencode_->isStop((int)i)?0:protAlphIndex_->getIndex(gencode_->translate((int)i));
     }
     
   };
